reorganizes file structure
