+++ include/Rendering/Texture.hpp
@@ -9,7 +9,7 @@
 #ifndef TEXTURE_HPP
 #define TEXTURE_HPP
 
-#include "Image.hpp"
+#include "Util/Image.hpp"
 
 #include <glad/glad.h>
 #include <string>
