+++ src/Rendering/Renderer.cpp
@@ -1,4 +1,4 @@
-#include "Renderer.hpp"
+#include "Rendering/Renderer.hpp"
 #include <iostream>
 
 // Sets the height and width of our renderer
