+++ src/Rendering/SkyBox.cpp
@@ -1,6 +1,6 @@
-#include "SkyBox.hpp"
+#include "Rendering/SkyBox.hpp"
 #define STB_IMAGE_IMPLEMENTATION
-#include "stb_image.h"
+#include "Util/stb_image.h"
 #include <glad/glad.h>
 #include <iostream>
